# docs/swagger.yaml
definitions:
...
  models.Category:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      name:
        type: string
      type:
        $ref: '#/definitions/models.CategoryType'
      updated_at:
        type: string
    type: object
...
  models.CategoryPagedResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/models.Category'
        type: array
      page_number:
        type: integer
      page_size:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  models.CategoryResponse:
    properties:
      data:
        $ref: '#/definitions/models.Category'
    type: object
  models.CategoryType:
    enum:
    - expense
    - income
    - transfer
    type: string
    x-enum-varnames:
    - CategoryTypeExpense
    - CategoryTypeIncome
    - CategoryTypeTransfer
...
  models.CreateCategoryRequest:
    properties:
      name:
        type: string
      type:
        allOf:
        - $ref: '#/definitions/models.CategoryType'
        enum:
        - expense
        - income
        - transfer
    required:
    - name
    - type
    type: object
...
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  models.UpdateCategoryRequest:
    properties:
      name:
        type: string
      type:
        $ref: '#/definitions/models.CategoryType'
    type: object
...
info:
  contact: {}
paths:
...
  /categories:
    get:
      consumes:
      - application/json
      description: Get a paginated list of categories with optional filtering
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      - description: Sort by field
        in: query
        name: sort_by
        type: string
      - description: Sort order (asc/desc)
        in: query
        name: sort_order
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CategoryPagedResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: List all categories
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: Create a new category with name and type
      parameters:
      - description: Category creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Create a new category
      tags:
      - categories
  /categories/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a category by its UUID
      parameters:
      - description: Category UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Delete a category
      tags:
      - categories
    get:
      consumes:
      - application/json
      description: Get a single category by its UUID
      parameters:
      - description: Category UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get category by ID
      tags:
      - categories
    put:
      consumes:
      - application/json
      description: Update an existing category by its UUID
      parameters:
      - description: Category UUID
        in: path
        name: id
        required: true
        type: string
      - description: Category update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Update a category
      tags:
      - categories
...
swagger: "2.0"